Small refactor to load storyblok in api route
